// src/models/staff-create-req.ts
// tslint:disable
/**
 * geo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 员工创建请求
 * @export
 * @interface StaffCreateReq
 */
export interface StaffCreateReq {
    /**
     * 用户名
     * @type {string}
     * @memberof StaffCreateReq
     */
    username: string;
    /**
     * 微信号
     * @type {string}
     * @memberof StaffCreateReq
     */
    wechat?: string;
    /**
     * 手机号
     * @type {string}
     * @memberof StaffCreateReq
     */
    phone: string;
    /**
     * 密码
     * @type {string}
     * @memberof StaffCreateReq
     */
    password: string;
    /**
     * 姓名
     * @type {string}
     * @memberof StaffCreateReq
     */
    name: string;
    /**
     * 角色
     * @type {string}
     * @memberof StaffCreateReq
     */
    role: StaffCreateReqRoleEnum;
    /**
     * 所属门店ID（队长/业务员必填）
     * @type {number}
     * @memberof StaffCreateReq
     */
    storeId?: number;
    /**
     * 所属队长ID（业务员必填）
     * @type {number}
     * @memberof StaffCreateReq
     */
    leaderId?: number;
    /**
     * 状态
     * @type {number}
     * @memberof StaffCreateReq
     */
    status?: number;
}

/**
    * @export
    * @enum {string}
    */
export enum StaffCreateReqRoleEnum {
    ADMIN = 'ADMIN',
    TEAMLEADER = 'TEAM_LEADER',
    SALESMAN = 'SALESMAN'
}